Standardized app.py and on_error messages in original samples.
