+++ samples/06.using-cards/requirements.txt
@@ -1,2 +1,3 @@
 botbuilder-core>=4.4.0b1
 botbuilder-dialogs>=4.4.0b1
+flask>=1.0.3
