+++ samples/13.core-bot/app.py
@@ -25,6 +25,7 @@ from bots import DialogAndWelcomeBot
 from adapter_with_error_handler import AdapterWithErrorHandler
 from flight_booking_recognizer import FlightBookingRecognizer
 
+# Create the loop and Flask app
 LOOP = asyncio.get_event_loop()
 APP = Flask(__name__, instance_relative_config=True)
 APP.config.from_object("config.DefaultConfig")
@@ -35,17 +36,22 @@ SETTINGS = BotFrameworkAdapterSettings(APP.config["APP_ID"], APP.config["APP_PAS
 MEMORY = MemoryStorage()
 USER_STATE = UserState(MEMORY)
 CONVERSATION_STATE = ConversationState(MEMORY)
+
+# Create adapter.
+# See https://aka.ms/about-bot-adapter to learn more about how bots work.
 ADAPTER = AdapterWithErrorHandler(SETTINGS, CONVERSATION_STATE)
+
+# Create dialogs and Bot
 RECOGNIZER = FlightBookingRecognizer(APP.config)
 BOOKING_DIALOG = BookingDialog()
-
 DIALOG = MainDialog(RECOGNIZER, BOOKING_DIALOG)
 BOT = DialogAndWelcomeBot(CONVERSATION_STATE, USER_STATE, DIALOG)
 
 
+# Listen for incoming requests on /api/messages.
 @APP.route("/api/messages", methods=["POST"])
 def messages():
-    """Main bot message handler."""
+    # Main bot message handler.
     if "application/json" in request.headers["Content-Type"]:
         body = request.json
     else:
@@ -56,12 +62,9 @@ def messages():
         request.headers["Authorization"] if "Authorization" in request.headers else ""
     )
 
-    async def aux_func(turn_context):
-        await BOT.on_turn(turn_context)
-
     try:
         task = LOOP.create_task(
-            ADAPTER.process_activity(activity, auth_header, aux_func)
+            ADAPTER.process_activity(activity, auth_header, BOT.on_turn)
         )
         LOOP.run_until_complete(task)
         return Response(status=201)
