+++ samples/45.state-management/config.py
@@ -7,13 +7,9 @@ import os
 """ Bot Configuration """
 
 
-class DefaultConfig(object):
+class DefaultConfig:
     """ Bot Configuration """
 
     PORT = 3978
     APP_ID = os.environ.get("MicrosoftAppId", "")
     APP_PASSWORD = os.environ.get("MicrosoftAppPassword", "")
-    LUIS_APP_ID = os.environ.get("LuisAppId", "")
-    LUIS_API_KEY = os.environ.get("LuisAPIKey", "")
-    # LUIS endpoint host name, ie "westus.api.cognitive.microsoft.com"
-    LUIS_API_HOST_NAME = os.environ.get("LuisAPIHostName", "")
