+++ samples/45.state-management/requirements.txt
@@ -1,1 +1,2 @@
 botbuilder-core>=4.4.0b1
+flask>=1.0.3
